requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
distro==1.8.0
frozenlist==1.3.3
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
magic-filter==1.0.9
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
Nuitka==1.5.5
ordered-set==4.1.0
peewee==3.16.0
Pygments==2.14.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rich==13.3.1
shell-gpt==0.8.5
sniffio==1.3.0
typer==0.7.0
urllib3==1.26.15
yarl==1.8.2
zstandard==0.20.0
